Fix T99191: Boolean modifier creates invalid material indices
Similar to 1a6d0ec71cf3b0c2c which changed the mesh boolean node (and
also caused this bug), this commit changes the material mapping for the
exact mode of the boolean modifier. Now the result should contain any
material on the faces of the input objects (including materials linked
to objects and meshes). The improvement is possible because materials
can be changed during evaluation (as of 1a81d268a19f2f1402).

Differential Revision: https://developer.blender.org/D15365